requirements.txt
fire
librosa
lightning
music21
pyMV2H
torchinfo
huggingface_hub
datasets
wandb
rich